+++ .gitmessage.txt
@@ -1,0 +1,34 @@
+# head (max 50 chars) -------------------------50^
+[lang] type: Message 
+# ex) [shell] feat: Add git action script
+# body (max 72 chars) -----------------------------------------------72^
+- 작업 내역 1
+- 작업 내역 2
+
+Refs:
+# 꼬릿말은 다음과 같이 작성: ex) Refs: #이슈 번호
+
+
+# --- COMMIT END ---
+# `[lang]` list
+#   python: Python으로 개발할 경우, e.g. [python]
+#   shell : Shell script의 수정만 존재할 경우, e.g. [shell]
+#   git   : README, .gitignore 등 git 관련 파일을 작성할 경우 [git]
+#   mixed: 여러 언어로 개발사항 또는 수정사항이 있을 경우, e.g. [mixed]
+#   etc : 특정 언어와 관련이 없을 경우, e.g. [etc]
+# `type:` list
+#   feat    : 기능 (새로운 기능)
+#   fix     : 버그 (버그 수정)
+#   refactor: 리팩토링
+#   style   : 스타일 (코드 형식, 세미콜론 추가: 비즈니스 로직에 변경 없음)
+#   docs    : 문서 (문서 추가, 수정, 삭제)
+#   test    : 테스트 (테스트 코드 추가, 수정, 삭제: 비즈니스 로직에 변경 없음)
+#   chore   : 기타 변경사항 (빌드 스크립트 수정 등)
+# ------------------
+#     제목 첫 글자를 대문자로
+#     제목은 명령문으로
+#     제목 끝에 마침표(.) 금지
+#     제목과 본문을 한 줄 띄워 분리하기
+#     본문은 "어떻게" 보다 "무엇을", "왜"를 설명한다.
+#     본문에 여러줄의 메시지를 작성할 땐 "-"로 구분
+# ------------------
