--- src/gpt2/__init__.py
@@ -1,1 +1,0 @@
-from .gpt2_dataset_builder import GPT2DatasetBuilder
