--- src/roberta/__init__.py
@@ -1,2 +1,0 @@
-from .roberta_dataset_builder import RobertaDatasetBuilder
-
