+++ train_tokenizer.py
@@ -1,9 +1,7 @@
 from dataclasses import dataclass, field
 from numpy.random import choice
-from src.dataset_loader import dataset_loader
-from src.utils import batch_iterator
+from src.utils import batch_iterator, load_corpora
 from transformers import AutoTokenizer, HfArgumentParser
-
 
 name_to_predefined_model = {
     "bert": "bert-base-uncased",
@@ -15,8 +13,8 @@ name_to_predefined_model = {
 
 @dataclass
 class DataArgument:
-    data_dir: str = field(
-        default = "corpora/wiki",
+    corpora_dir: str = field(
+        default = "corpora/kowiki",
     )
     batch_size: int = field(
         default = 1000,
@@ -42,24 +40,24 @@ class ModelArgument:
     vocab_size: int = field(
         default = 30000,
     )
-    model_max_length: int = field(
-        default = 512,
+    min_frequency: int = field(
+        default = 2,
     )
 
 
 def main():
     parser = HfArgumentParser((DataArgument, ModelArgument))
     data_args, model_args = parser.parse_args_into_dataclasses()
-    dataset = dataset_loader(data_args.data_dir)
+    corpora = load_corpora(data_args.corpora_dir)
 
     if 0 < data_args.sampling_ratio < 1.0:
-        total_size = len(dataset)
+        total_size = len(corpora)
         sample_size = int(total_size * data_args.sampling_ratio)
-        dataset = dataset.select(indices=choice(range(total_size), sample_size))
+        sampled_corpora = corpora.select(indices=choice(range(total_size), sample_size))
 
     tokenizer = AutoTokenizer.from_pretrained(name_to_predefined_model[model_args.model_name])
-    data_iterator = batch_iterator(dataset, batch_size=data_args.batch_size)
-    tokenizer = tokenizer.train_new_from_iterator(data_iterator, vocab_size=model_args.vocab_size)
+    data_iterator = batch_iterator(sampled_corpora, batch_size=data_args.batch_size)
+    tokenizer = tokenizer.train_new_from_iterator(data_iterator, vocab_size=model_args.vocab_size, min_frequency=model_args.min_frequency)
     tokenizer.save_pretrained("tokenizers/" + model_args.model_name)
